task_1/requirements.txt
# Required packages for LoRA fine-tuning and inference
transformers>=4.36.0
peft>=0.7.0
torch>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.62.0
requests>=2.27.0
tensorboard>=2.10.0
accelerate>=0.21.0
bitsandbytes>=0.39.0
safetensors>=0.3.0
sentencepiece>=0.1.97
protobuf>=3.20.0
datasets>=2.10.0
python-dotenv>=1.0.0

